requirements.txt
django==2.2.8
django-autocomplete-light
django-cities-light
django-import-export
pandas
Pillow